requirements.txt
streamlit
psutil
python-dotenv
pytest